+++ src/components/MyNyplButton/MyNyplButton.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import { extend as _extend } from 'underscore';
 import FocusTrap from 'focus-trap-react';
 import {
-  DownWedgeIcon,
+  GenericWedgeIcon,
   XIcon,
-} from 'dgx-svg-icons';
+} from '@nypl/dgx-svg-icons';
 // GA Utilities
 import utils from '../../utils/utils.js';
 // Component Dependencies
@@ -36,6 +36,7 @@ const styles = {
     minWidth: '250px',
     backgroundColor: '#1B7FA7',
     padding: '25px 30px',
+    marginTop: '10px',
   },
 };
 
@@ -102,7 +103,7 @@ class MyNyplButton extends React.Component {
    */
   renderMyNyplButton() {
     let buttonClass = '';
-    let icon = <DownWedgeIcon className="dropDownIcon" ariaHidden />;
+    let icon = <GenericWedgeIcon className="dropDownIcon" ariaHidden />;
     let myNyplButtonLabel = (this.props.patronName) ? 'My Account' : 'Log In';
     const labelColorClass = (this.props.isLoggedIn) ? ' loggedIn' : '';
     const loggedInFadeInAnimation = (this.props.patronName) ? ' animated fadeIn' : '';
@@ -115,7 +116,7 @@ class MyNyplButton extends React.Component {
 
     return (
       <a
-        className={`MyNyplButton ${buttonClass}${labelColorClass}${loggedInFadeInAnimation}`}
+        className={`myNyplButton ${buttonClass}${labelColorClass}${loggedInFadeInAnimation}`}
         onClick={this.handleClick}
         style={_extend(styles.MyNyplButton, this.props.style)}
         href={this.props.target}
@@ -133,7 +134,7 @@ class MyNyplButton extends React.Component {
     const boxHeight = (this.props.isLoggedIn) ? ' loggedInHeight' : null;
     return (this.state.visible) ? (
       <div
-        className={`MyNypl-Wrapper active animatedFast fadeIn${boxHeight}`}
+        className={`myNypl-wrapper active animatedFast fadeIn${boxHeight}`}
         style={styles.MyNyplWrapper}
       >
         <MyNypl
@@ -153,7 +154,7 @@ class MyNyplButton extends React.Component {
           clickOutsideDeactivates: true,
         }}
         active={this.state.visible}
-        className="MyNyplButton-Wrapper"
+        className="myNyplButton-wrapper"
         style={_extend(styles.base, this.props.style)}
       >
         {this.renderMyNyplButton()}