+++ src/components/NavMenuMobileButtons/NavMenuMobileButtons.js
@@ -1,6 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { extend as _extend } from 'underscore';
+import {
+  LibraryCardIcon,
+  EnvelopeIcon,
+  XIcon,
+} from '@nypl/dgx-svg-icons';
+
 import utils from '../../utils/utils.js';
 // Dependent NYPL React Component
 import DonateButton from '../DonateButton/DonateButton.js';
@@ -70,8 +76,8 @@ const NavMenuMobileButtons = ({
   libraryCardLink,
   subscribeLink,
 }) => {
-  const libraryCardClass = 'LibraryCardLink';
-  const subscribeLinkClass = 'SubscribeLink';
+  const libraryCardClass = 'libraryCardLink';
+  const subscribeLinkClass = 'subscribeLink';
 
   return (
     <div className={className} style={styles.base}>
@@ -82,17 +88,12 @@ const NavMenuMobileButtons = ({
         onClick={() => utils.trackHeader('Click', 'Mobile Bottom Buttons - Library Card')}
       >
         <span
-          className={`${libraryCardClass}-Wrapper`}
+          className={`${libraryCardClass}-wrapper`}
           style={_extend(styles.wrapper, styles.libraryCardLinkWrapper)}
         >
+          <LibraryCardIcon iconId="libraryCardSVG" ariaHidden />
           <span
-            className={`${libraryCardClass}-Icon nypl-icon-card`}
-            style={styles.icon}
-            aria-hidden="true"
-          >
-          </span>
-          <span
-            className={`${libraryCardClass}-Label`}
+            className={`${libraryCardClass}-label`}
             style={_extend(styles.libraryCardLinkLabel, styles.label)}
           >
             Get a Library Card
@@ -106,17 +107,12 @@ const NavMenuMobileButtons = ({
         onClick={() => utils.trackHeader('Click', 'Mobile Bottom Buttons - Email Updates')}
       >
         <span
-          className={`${subscribeLinkClass}-Wrapper`}
+          className={`${subscribeLinkClass}-wrapper`}
           style={_extend(styles.wrapper, styles.subscribeLinkWrapper)}
         >
+          <EnvelopeIcon iconId="envelopSVG" ariaHidden />
           <span
-            className={`${subscribeLinkClass}-Icon nypl-icon-mail`}
-            style={styles.icon}
-            aria-hidden="true"
-          >
-          </span>
-          <span
-            className={`${subscribeLinkClass}-Label`}
+            className={`${subscribeLinkClass}-label`}
             style={_extend(styles.subscribeLinkLabel, styles.label)}
           >
             Get Email Updates
@@ -124,8 +120,8 @@ const NavMenuMobileButtons = ({
         </span>
       </a>
       <DonateButton
-        id="MobileNav-DonateButton"
-        className="DonateLink"
+        id="mobileNav-donateButton"
+        className="donateLink"
         style={styles.donateLink}
         gaLabel="Mobile Buttons Donate"
       />
@@ -142,9 +138,9 @@ NavMenuMobileButtons.propTypes = {
 
 NavMenuMobileButtons.defaultProps = {
   lang: 'en',
-  className: 'NavMenuMobileButtons',
+  className: 'navMenuMobileButtons',
   libraryCardLink: '//www.nypl.org/library-card',
-  subscribeLink: '//pages.email.nypl.org/page.aspx' +
+  subscribeLink: 'http://pages.email.nypl.org/page.aspx' +
     '?QS=3935619f7de112ef7250fe02b84fb2f9ab74e4ea015814b7',
 };